requirements.txt
Click==7.0
Flask==1.1.2
Flask_APScheduler==1.12.2
requests==2.21.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
Flask_SQLAlchemy==2.5.1
PyJWT==2.1.0
redis==3.5.3
rq==1.10.0
